Revert "Revert "NDRS-925 block executor optimization (1/2)""
This reverts my reverted commit 83f8d01fca26ba8af61448b6485a53e45c78b647.
