+++ node/src/components/gossiper/tests.rs
@@ -8,6 +8,7 @@ use std::{
 use derive_more::From;
 use prometheus::Registry;
 use rand::Rng;
+use semver::Version;
 use serde::Serialize;
 use tempfile::TempDir;
 use thiserror::Error;
@@ -22,12 +23,13 @@ use crate::{
         in_memory_network::{self, InMemoryNetwork, NetworkController},
         storage::{self, Storage},
     },
+    crypto::hash::Digest,
     effect::{
         announcements::{
-            DeployAcceptorAnnouncement, GossiperAnnouncement, NetworkAnnouncement,
-            RpcServerAnnouncement,
+            ContractRuntimeAnnouncement, DeployAcceptorAnnouncement, GossiperAnnouncement,
+            NetworkAnnouncement, RpcServerAnnouncement,
         },
-        requests::ContractRuntimeRequest,
+        requests::{ConsensusRequest, ContractRuntimeRequest, LinearChainRequest},
         Responder,
     },
     protocol::Message as NodeMessage,
@@ -75,13 +77,31 @@ impl From<StorageRequest> for Event {
 
 impl From<ContractRuntimeRequest> for Event {
     fn from(request: ContractRuntimeRequest) -> Self {
-        Event::ContractRuntime(contract_runtime::Event::Request(request))
+        Event::ContractRuntime(contract_runtime::Event::Request(Box::new(request)))
     }
 }
 
 impl From<NetworkRequest<NodeId, Message<Deploy>>> for Event {
     fn from(request: NetworkRequest<NodeId, Message<Deploy>>) -> Self {
         Event::NetworkRequest(request.map_payload(NodeMessage::from))
+    }
+}
+
+impl From<ConsensusRequest> for Event {
+    fn from(_request: ConsensusRequest) -> Self {
+        unimplemented!("not implemented for gossiper tests")
+    }
+}
+
+impl From<LinearChainRequest<NodeId>> for Event {
+    fn from(_request: LinearChainRequest<NodeId>) -> Self {
+        unimplemented!("not implemented for gossiper tests")
+    }
+}
+
+impl From<ContractRuntimeAnnouncement> for Event {
+    fn from(_request: ContractRuntimeAnnouncement) -> Self {
+        unimplemented!("not implemented for gossiper tests")
     }
 }
 
@@ -104,7 +124,7 @@ impl Display for Event {
                 write!(formatter, "deploy-gossiper announcement: {}", ann)
             }
             Event::ContractRuntime(event) => {
-                write!(formatter, "contract-runtime event: {}", event)
+                write!(formatter, "contract-runtime event: {:?}", event)
             }
         }
     }
@@ -150,8 +170,15 @@ impl reactor::Reactor for Reactor {
         let storage = Storage::new(&storage_withdir, None).unwrap();
 
         let contract_runtime_config = contract_runtime::Config::default();
-        let contract_runtime =
-            ContractRuntime::new(storage_withdir, &contract_runtime_config, &registry).unwrap();
+        let contract_runtime = ContractRuntime::new(
+            Digest::random(rng),
+            None,
+            Version::new(1, 0, 0),
+            storage_withdir,
+            &contract_runtime_config,
+            &registry,
+        )
+        .unwrap();
 
         let deploy_acceptor = DeployAcceptor::new(
             deploy_acceptor::Config::new(false),
