+++ node/src/effect/announcements.rs
@@ -174,23 +174,44 @@ where
     }
 }
 
-/// A BlockExecutor announcement.
+/// A ContractRuntime announcement.
 #[derive(Debug)]
-pub enum BlockExecutorAnnouncement {
+pub enum ContractRuntimeAnnouncement {
     /// A new block from the linear chain was produced.
-    LinearChainBlock {
-        /// The block.
-        block: Block,
-        /// The results of executing the deploys in this block.
-        execution_results: HashMap<DeployHash, (DeployHeader, ExecutionResult)>,
-    },
+    LinearChainBlock(Box<LinearChainBlock>),
 }
 
-impl Display for BlockExecutorAnnouncement {
+impl ContractRuntimeAnnouncement {
+    /// Create a ContractRuntimeAnnouncement::LinearChainBlock from it's parts.
+    pub fn linear_chain_block(
+        block: Block,
+        execution_results: HashMap<DeployHash, (DeployHeader, ExecutionResult)>,
+    ) -> Self {
+        Self::LinearChainBlock(Box::new(LinearChainBlock {
+            block,
+            execution_results,
+        }))
+    }
+}
+
+/// A ContractRuntimeAnnouncement's block.
+#[derive(Debug)]
+pub struct LinearChainBlock {
+    /// The block.
+    pub block: Block,
+    /// The results of executing the deploys in this block.
+    pub execution_results: HashMap<DeployHash, (DeployHeader, ExecutionResult)>,
+}
+
+impl Display for ContractRuntimeAnnouncement {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
-            BlockExecutorAnnouncement::LinearChainBlock { block, .. } => {
-                write!(f, "created linear chain block {}", block.hash())
+            ContractRuntimeAnnouncement::LinearChainBlock(linear_chain_block) => {
+                write!(
+                    f,
+                    "created linear chain block {}",
+                    linear_chain_block.block.hash()
+                )
             }
         }
     }
