+++ node/src/reactor/joiner/memory_metrics.rs
@@ -34,8 +34,6 @@ pub struct MemoryMetrics {
     mem_block_validator: IntGauge,
     /// Estimated heap memory usage of deploy fetcher component.
     mem_deploy_fetcher: IntGauge,
-    /// Estimated heap memory usage of block executor component.
-    mem_block_executor: IntGauge,
     /// Estimated heap memory usage of linear chain component.
     mem_linear_chain: IntGauge,
     /// Estimated heap memory usage of consensus component.
@@ -88,8 +86,6 @@ impl MemoryMetrics {
             "joiner_mem_deploy_fetcher",
             "deploy_fetcher memory usage in bytes",
         )?;
-        let mem_block_executor =
-            IntGauge::new("mem_block_executor", "block_executor memory usage in bytes")?;
         let mem_linear_chain = IntGauge::new(
             "joiner_mem_linear_chain",
             "linear_chain memory usage in bytes",
@@ -118,7 +114,6 @@ impl MemoryMetrics {
         registry.register(Box::new(mem_linear_chain_sync.clone()))?;
         registry.register(Box::new(mem_block_validator.clone()))?;
         registry.register(Box::new(mem_deploy_fetcher.clone()))?;
-        registry.register(Box::new(mem_block_executor.clone()))?;
         registry.register(Box::new(mem_linear_chain.clone()))?;
         registry.register(Box::new(mem_consensus.clone()))?;
 
@@ -136,7 +131,6 @@ impl MemoryMetrics {
             mem_linear_chain_sync,
             mem_block_validator,
             mem_deploy_fetcher,
-            mem_block_executor,
             mem_linear_chain,
             mem_consensus,
             mem_estimator_runtime_s,
@@ -160,7 +154,6 @@ impl MemoryMetrics {
         let linear_chain_sync = reactor.linear_chain_sync.estimate_heap_size() as i64;
         let block_validator = reactor.block_validator.estimate_heap_size() as i64;
         let deploy_fetcher = reactor.deploy_fetcher.estimate_heap_size() as i64;
-        let block_executor = reactor.block_executor.estimate_heap_size() as i64;
         let linear_chain = reactor.linear_chain.estimate_heap_size() as i64;
         let consensus = reactor.consensus.estimate_heap_size() as i64;
 
@@ -176,7 +169,6 @@ impl MemoryMetrics {
             + linear_chain_sync
             + block_validator
             + deploy_fetcher
-            + block_executor
             + linear_chain
             + consensus;
 
@@ -193,7 +185,6 @@ impl MemoryMetrics {
         self.mem_linear_chain_sync.set(linear_chain_sync);
         self.mem_block_validator.set(block_validator);
         self.mem_deploy_fetcher.set(deploy_fetcher);
-        self.mem_block_executor.set(block_executor);
         self.mem_linear_chain.set(linear_chain);
         self.mem_consensus.set(consensus);
 
@@ -215,7 +206,6 @@ impl MemoryMetrics {
         %linear_chain_sync,
         %block_validator,
         %deploy_fetcher,
-        %block_executor,
         %linear_chain,
         %consensus,
         "Collected new set of memory metrics for the joiner");
@@ -274,11 +264,6 @@ impl Drop for MemoryMetrics {
             .unregister(Box::new(self.mem_deploy_fetcher.clone()))
             .unwrap_or_else(
                 |err| warn!(%err, "did not expect deregistering joiner_mem_deploy_fetcher to fail"),
-            );
-        self.registry
-            .unregister(Box::new(self.mem_block_executor.clone()))
-            .unwrap_or_else(
-                |err| warn!(%err, "did not expect deregistering joiner_mem_block_executor to fail"),
             );
         self.registry
             .unregister(Box::new(self.mem_linear_chain.clone()))
